Fixed login form design
