+++ app/src/main/AndroidManifest.xml
@@ -27,7 +27,7 @@
             android:name=".RegisterActivity"
             android:exported="false" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".ui.auth.LoginActivity"
             android:exported="false" />
         <!-- SECURITY: Token received via deep link fragment must be verified server-side before use -->
         <activity
