+++ app/src/main/java/com/example/ucms_android/session/SessionManager.java
@@ -1,5 +1,43 @@
 package com.example.ucms_android.session;
 
+import android.content.Context;
+import android.content.SharedPreferences;
+
 public class SessionManager {
-    // TODO: JWT + role storage via SharedPreferences
+    private static final String PREF_NAME = "ucms_session";
+    private static final String KEY_TOKEN = "access_token";
+    private static final String KEY_ROLE = "user_role";
+    private static final String KEY_LOGGED_IN = "is_logged_in";
+
+    private final SharedPreferences sharedPreferences;
+    private final SharedPreferences.Editor editor;
+
+    public SessionManager(Context context) {
+        sharedPreferences = context.getSharedPreferences(PREF_NAME, Context.MODE_PRIVATE);
+        editor = sharedPreferences.edit();
+    }
+
+    public void saveSession(String token, String role) {
+        editor.putString(KEY_TOKEN, token);
+        editor.putString(KEY_ROLE, role);
+        editor.putBoolean(KEY_LOGGED_IN, true);
+        editor.apply();
+    }
+
+    public String getToken() {
+        return sharedPreferences.getString(KEY_TOKEN, null);
+    }
+
+    public String getRole() {
+        return sharedPreferences.getString(KEY_ROLE, "STUDENT");
+    }
+
+    public boolean isLoggedIn() {
+        return sharedPreferences.getBoolean(KEY_LOGGED_IN, false);
+    }
+
+    public void clearSession() {
+        editor.clear();
+        editor.apply();
+    }
 }
