+++ app/src/main/res/drawable/bg_button_pill.xml
@@ -2,7 +2,7 @@
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle">
 
-    <solid android:color="@color/colorPrimary" />
+    <solid android:color="@color/colorSecondary" />
 
     <corners android:radius="@dimen/corner_radius_button" />
