+++ app/src/main/res/drawable/bg_input.xml
@@ -2,12 +2,12 @@
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle">
 
-    <solid android:color="@color/colorBackground" />
+    <solid android:color="@color/white" />
 
     <stroke
-        android:width="@dimen/stroke_width_default"
-        android:color="@color/colorDivider" />
+        android:width="1dp"
+        android:color="#E0E0E0" />
 
-    <corners android:radius="@dimen/corner_radius_input" />
+    <corners android:radius="16dp" />
 
 </shape>
