+++ app/src/main/res/values/strings.xml
@@ -3,6 +3,7 @@
     <string name="login">Login</string>
     <string name="register">Register</string>
     <string name="student_id">Student ID</string>
+    <string name="student_id_label">Student-ID</string>
     <string name="password">Password</string>
     <string name="full_name">Full Name</string>
     <string name="course">Course</string>
